move more global configuration
